ptrcreate.sh - Building a PTR Record
In this screen cast I will be showing how to create a PTR record for a Rackspace NextGen Cloud Server using a script that consumes the API.

This is a simple script that will create a PTR record for a Next Generation Cloud Server.

The function of the script is such that, you simply need to enter your details and all of the rest is done automagically.

Use Case :
Create a PTR (Reverse DNS) Record
Update a PTR (Reverse DNS) Record

Application :
The script can be used with all variables passed on the command line or nteractively.

To get the script you can go to my github page found here :
https://github.com/cloudnull/ptrcreate ( https://github.com/cloudnull/ptrcreate )