GeneAnnoLogy demo: 20 Aug 2012
I've implemented a lot of the core functionality that I foresee being important for GeneAnnoLogy. There is still a lot of work required to improve and complete the program, but it's definitely ready to be put through some tests. Here I show the basic usage with a toy data set from Arabidopsis thaliana .